FIXED: Possible deadlock when running the development tools in a thread. Jacco van Ossenbruggen.
This might not be the last word on this. The problem is that the main
thread interacts with XPCE in such a way that it requires a reply from
the X-server, but this feedback is handed to the pce thread that
processed events. This can be a more general problem in X11 interaction,
but it may also be a race condition when starting the development
environment.  We'll see ...