FIXED: Avoid crash when trying to redraw a window that is not associated to a display.  Jochem Liem.
It should be impossible to get into this situation, but appearently this
is possible.  It is hard to figure out whether this is caused by illegal
actions on XPCE objects or by missing code that ensures that window<-displayed
is always @off if a window is not associated with a frame.