ENHANCED: Graphical debugger retry on skipped goals.
Initially, the exit of a skipped goal is answered with `creep' to get to
the caller.  This is not implemented on top of the new skipped attribute,
such that it also works for retried skipped goals.  Not sure whether the
rest of the condition still makes sense (but it probably won't hurt).