ENHANCED: Avoid blocking on messages sent to XPCE.
When using in_pce_thread(), messages are sent to a pipe.  If the XPCE
thread itself is not processing messages, this results in a deadlock
if the pipe is full.

The work-around is partial, limiting messages sent to xpce and
avoid the use of in_pce_thread() if we are the xpce thread.