PORT: Make the xpce source compile clean using GCC -pedantic flag.
Note that an important part of the problems is that XPCE threads data and function
pointers the same.   This used to work with a cast to `void*`.  This is not allowed
in C11.  It is allowed to cast both pointers and functions to `intptr_t`, so with
a little detour we get where we need to be.  Time will tell whether we will see
systems where this conversion is not allowed.