MODIFIED: WM_CLASS handling
Recent desktop environments such as Gnome 46 and later ignore
the ``_NET_WM_ICON`` property to set an icon for a frame
(toplevel window).  We can set the icon through a ``.desktop``
file using ``StartupWMClass`` where all windows of a process
are bundled under the same icon.   But, xpce set ``WM_CLASS``
from the xpce class name.

This patch adds `display<->wm_class` to set the ``WM_CLASS``
for all toplevel windows.  The default is ``SWI-Prolog``.  By
setting this to ``@default``, the old behavior is restored.