Make menu_bar and popup cooperate without grabbing the mouse.
In the old design, all events go through the menu_bar.   These are passed
to the popups.  Rather than executing, the popup closes and sets the
current_item.   The menubar detects this, releases the focus and executes
the popup.

In Wayland, grabbing the pointer is not (always) allowed.  So, we make
the popup execute the action.