FIXED: Make Window creation display aware.
On Wayland, new window positions are ignored, but Wayland creates new windows on
the "active" display.  On MacOS, windows are always created on the primary display
unless we place them in the area of a specific display.   Our code now tracks the
"active" display and opens new windows on this display.