ENHANCED: Capture xpce messages in the epilog console.
This patch redirects direct console output and Prolog print_message/2
that would print to the Prolog console to an Epilog window.   If the
message is associated to a thread with an epilog window, use that.  Else
use the last used Epilog window or just any of them as last resort.